dev_requirements.txt
appdirs==1.4.4
blosc
pycairo
certifi==2022.12.7
click==8.1.3
cryptography==39.0.0
cycler==0.11.0
fiona==1.8.22
geopandas==0.12.2
geos
h5py
jinja2==3.1.2
kiwisolver==1.4.4
matplotlib
numpy==1.21.4
sqlalchemy==1.4.41
